Move AssetChangeTextUi4->text/AssetChangeText
